move yaml utils to module_utils
