+++ lib/ansible/galaxy/collection/__init__.py
@@ -117,11 +117,11 @@ from ansible.galaxy.dependency_resolution.errors import (
 from ansible.galaxy.dependency_resolution.versioning import meets_requirements
 from ansible.module_utils.six import raise_from
 from ansible.module_utils._text import to_bytes, to_native, to_text
+from ansible.module_utils.common.yaml import yaml_dump
 from ansible.utils.collection_loader import AnsibleCollectionRef
 from ansible.utils.display import Display
 from ansible.utils.hashing import secure_hash, secure_hash_s
 from ansible.utils.version import SemanticVersion
-from ansible.utils.yaml import yaml_dump
 
 
 display = Display()
