+++ lib/ansible/galaxy/role.py
@@ -34,10 +34,10 @@ from ansible import context
 from ansible.errors import AnsibleError
 from ansible.galaxy.user_agent import user_agent
 from ansible.module_utils._text import to_native, to_text
+from ansible.module_utils.common.yaml import yaml_dump, yaml_load
 from ansible.module_utils.urls import open_url
 from ansible.playbook.role.requirement import RoleRequirement
 from ansible.utils.display import Display
-from ansible.utils.yaml import yaml_dump, yaml_load
 
 display = Display()
